requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
sqlite-utils==3.36
colorlog==6.8.2
